# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Enershare - REC Operation and LEM pricing API",
        "description": "A REST API designed to calculate the prices in a Local Energy Market (LEM) and to determine the operational schedule of Battery Energy Storage System (BESS) assets within a Renewable Energy Community (REC).",
        "version": "0.2.1"
    },
    "paths": {
        "/vanilla/{pricing_mechanism}": {
            "post": {
                "tags": [
                    "Calculate LEM Prices"
                ],
                "summary": "Vanilla",
                "description": "Calculate an array of LEM prices using the selected pricing mechanism. <br />No Mixed Integer Linear Programming (MILP) is solved. The LEM offers are formulated solely on the basis of the meters’ historical or projected net consumption and their corresponding opportunity costs",
                "operationId": "vanilla_vanilla__pricing_mechanism__post",
                "parameters": [
                    {
                        "name": "pricing_mechanism",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/PricingMechanism"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserParams"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AcceptedResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/vanilla/{order_id}": {
            "get": {
                "tags": [
                    "Retrieve LEM Prices"
                ],
                "summary": "Vanilla",
                "description": "An endpoint for fetching the results of a \"vanilla\" request, given the order ID.",
                "operationId": "vanilla_vanilla__order_id__get",
                "parameters": [
                    {
                        "name": "order_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Order Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VanillaOutputs"
                                }
                            }
                        }
                    },
                    "202": {
                        "description": "Order found but not yet processed.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderNotProcessed"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Order not found.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderNotFound"
                                }
                            }
                        }
                    },
                    "412": {
                        "description": "One or more meter IDs not found.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MeterIDNotFound"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "One or more data points for one or more meter IDs not found.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TimeseriesDataNotFound"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/dual": {
            "post": {
                "tags": [
                    "Schedule operation and calculate LEM Prices"
                ],
                "summary": "Dual",
                "description": "Calculate an array of LEM prices and the operational schedule of the BESS assets by executing a purely collective MILP. <br />In this process, the shadow prices of a LEM equilibrium constraint are returned as the optimal LEM prices.",
                "operationId": "dual_dual_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BaseUserParams"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "202": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AcceptedResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/loop/{lem_organization}/{pricing_mechanism}": {
            "post": {
                "tags": [
                    "Schedule operation and calculate LEM Prices"
                ],
                "summary": "Loop",
                "description": "Calculate an array of LEM prices and the operational schedule of the BESS assets by implementing an iterative algorithm. <br />In this process, successive two-stage MILP are solved until a specified stopping criterion is achieved. Each two-stage MILP procedure is executed with LEM prices that are calculated based on the user-defined pricing mechanism. The offers are formulated based on the net loads that result from the previously solved MILP.",
                "operationId": "loop_loop__lem_organization___pricing_mechanism__post",
                "parameters": [
                    {
                        "name": "pricing_mechanism",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/PricingMechanism"
                        }
                    },
                    {
                        "name": "lem_organization",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/LemOrganization"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserParams"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AcceptedResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/dual/{order_id}": {
            "get": {
                "tags": [
                    "Retrieve operation and LEM prices"
                ],
                "summary": "Dual",
                "description": "An endpoint for fetching the results of a request that involves solving one or several MILP, given the order ID.",
                "operationId": "dual_dual__order_id__get",
                "parameters": [
                    {
                        "name": "order_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Order Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PoolMILPOutputs"
                                }
                            }
                        }
                    },
                    "202": {
                        "description": "Order found but not yet processed.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderNotProcessed"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Order not found.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderNotFound"
                                }
                            }
                        }
                    },
                    "412": {
                        "description": "One or more meter IDs not found.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MeterIDNotFound"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "One or more data point for one or more meter IDs not found.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TimeseriesDataNotFound"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "AcceptedResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "title": "Message",
                        "examples": [
                            "Processing has started. Use the order ID for status updates."
                        ]
                    },
                    "order_id": {
                        "type": "string",
                        "title": "Order Id",
                        "description": "Order identifier for the request. <br />Request results via REST API can only be retrieved by specifying this identifier.",
                        "examples": [
                            "iaMiULXA9BktPUu2b_PwTtycCSNe0_wYpPt9muwlEtgL49GDg-kggSktAjtu"
                        ]
                    }
                },
                "type": "object",
                "required": [
                    "message",
                    "order_id"
                ],
                "title": "AcceptedResponse"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "LemOrganization": {
                "type": "string",
                "enum": [
                    "pool",
                    "bilateral"
                ],
                "title": "LemOrganization"
            },
...
            "PricingMechanism": {
                "type": "string",
                "enum": [
                    "crossing_value",
                    "mmr",
                    "sdr"
                ],
                "title": "PricingMechanism"
            },
...
            "VanillaOutputs": {
                "properties": {
                    "order_id": {
                        "type": "string",
                        "maxLength": 60,
                        "minLength": 60,
                        "title": "Order Id",
                        "description": "Order identifier for the request.",
                        "examples": [
                            "iaMiULXA9BktPUu2b_PwTtycCSNe0_wYpPt9muwlEtgL49GDg-kggSktAjtu"
                        ]
                    },
                    "lem_prices": {
                        "items": {
                            "$ref": "#/components/schemas/LemPrice"
                        },
                        "type": "array",
                        "title": "Lem Prices",
                        "description": "List with the local energy market prices computed for the requested horizon."
                    },
                    "offers": {
                        "items": {
                            "$ref": "#/components/schemas/Offer"
                        },
                        "type": "array",
                        "title": "Offers",
                        "description": "A description of all the offers considered in the local energy market session that resulted in the calculated price."
                    }
                },
                "type": "object",
                "required": [
                    "order_id",
                    "lem_prices",
                    "offers"
                ],
                "title": "VanillaOutputs"
            }
        }
    }
}